Fixes docking space alignment bug
Bug occurs when the window is moved out of the top left corner of the screen. When the window of the application moves but the docking space stays in the screen's global location (0, 0). By setting the next window pos to the pos of the application it makes sure the content stays in view.